+++ CHANGELOG.md
@@ -45,3 +45,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Enhance Securial controllers and specs with error handling and authorization checks
 - Implement authentication and authorization mechanisms in Securial controllers
 - implement authenticate_admin! for actions that require admin/superuser role
+- Implement sign-in, sign-out, refresh token, revoke token functionality